+++ src/hotspot/os/windows/os_windows.hpp
@@ -133,57 +133,4 @@ public:
   static inline int get_thread_ptr_offset() { return _thread_ptr_offset; }
 };
 
-/*
- * Crash protection for the JfrSampler thread. Wrap the callback
- * with a __try { call() }
- * To be able to use this - don't take locks, don't rely on destructors,
- * don't make OS library calls, don't allocate memory, don't print,
- * don't call code that could leave the heap / memory in an inconsistent state,
- * or anything else where we are not in control if we suddenly jump out.
- */
-class ThreadCrashProtection : public StackObj {
-public:
-  static bool is_crash_protected(Thread* thr) {
-    return _crash_protection != NULL && _protected_thread == thr;
-  }
-
-  ThreadCrashProtection();
-  bool call(os::CrashProtectionCallback& cb);
-private:
-  static Thread* _protected_thread;
-  static ThreadCrashProtection* _crash_protection;
-};
-
-// Platform specific implementations that underpin VM Mutex/Monitor classes.
-// Note that CRITICAL_SECTION supports recursive locking, while the semantics
-// of the VM Mutex class does not. It is up to the Mutex class to hide this
-// difference in behaviour.
-
-class PlatformMutex : public CHeapObj<mtSynchronizer> {
-  NONCOPYABLE(PlatformMutex);
-
- protected:
-  CRITICAL_SECTION   _mutex; // Native mutex for locking
-
- public:
-  PlatformMutex();
-  ~PlatformMutex();
-  void lock();
-  void unlock();
-  bool try_lock();
-};
-
-class PlatformMonitor : public PlatformMutex {
- private:
-  CONDITION_VARIABLE _cond;  // Native condition variable for blocking
-  NONCOPYABLE(PlatformMonitor);
-
- public:
-  PlatformMonitor();
-  ~PlatformMonitor();
-  int wait(jlong millis);
-  void notify();
-  void notify_all();
-};
-
 #endif // OS_WINDOWS_OS_WINDOWS_HPP
